Correct hex_coverage table

+++ mobile_verifier/migrations/13_modeled_coverage.sql
@@ -7,7 +7,7 @@ CREATE TYPE signal_level as ENUM (
 
 CREATE TABLE hex_coverage (
        uuid UUID NOT NULL,
-       hex INTEGER NOT NULL,
+       hex BIGINT NOT NULL,
        indoor BOOLEAN NOT NULL,
        cbsd_id TEXT NOT NULL,
        signal_level signal_level NOT NULL,
